Remove redundant folders and modularize notifications
- Delete unused app/BERT/ (no production consumers).
- Delete app/utils/scraper.py shim (just re-exported app.searches.execution).
- Move app/utils/notifications.py into app/notifications/telegram.py so
  Telegram transport lives alongside the rest of the notifications
  domain (service/events/relevance/delivery/records/preferences).
- Update all consumers and patch paths in tests.
