+++ app/notifications/telegram.py
@@ -1,6 +1,7 @@
 import requests
-from flask import current_app, url_for
+from flask import current_app
 from datetime import datetime, timezone
+
 
 class NotificationHandler:
     @staticmethod
@@ -10,11 +11,12 @@ class NotificationHandler:
             user.notification_preferences.get(notification_type, True)
         )
 
+
 class TelegramNotifier:
     def __init__(self, bot_token, chat_id):
         self.bot_token = bot_token
         self.chat_id = chat_id
-        
+
     def send_message(self, message):
         """Send formatted message through Telegram"""
         try:
@@ -34,21 +36,18 @@ class TelegramNotifier:
             current_app.logger.error(f"Telegram send failed: {str(e)}")
             return False
 
+
 class NotificationManager:
     @staticmethod
     def send_item_notification(user, items, query_text=None):
-        # Check preferences and connection
         print("Send item notification running")
-        if (not user.telegram_connected or 
+        if (not user.telegram_connected or
             not user.notification_preferences.get('new_items', True) or
             not items):
             return False
-        
+
         print("Main telegram ID: ", user.telegram_chat_ids['main'])
-        
-        
-        
-        
+
         try:
             chat_ids = [user.telegram_chat_ids['main']] + user.telegram_chat_ids['additional']
             for chat_id in chat_ids:
@@ -57,34 +56,29 @@ class NotificationManager:
                     chat_id
                 )
                 print("Chat ID: ", chat_ids)
-            
-                # Improved message formatting
+
                 query_text = f" for your '{query_text}' search" if query_text else ""
                 message = (
-                f"🎉 <b>New Items Found{query_text}!</b>\n\n"
-                f"📥 Total new items: {len(items)}\n\n"
-            )
-            
-                # Add top 5 items
+                    f"🎉 <b>New Items Found{query_text}!</b>\n\n"
+                    f"📥 Total new items: {len(items)}\n\n"
+                )
+
                 for item in items[:5]:
                     message += (
-                    f"🏷️ <a href='{item.url}'>{item.title}</a>\n"
-                    f"💰 Price: {item.price} {item.currency}\n"
-                    f"📍 Location: {item.location_country or 'N/A'}\n\n"
-                )
-                    
-                
-            
-                # Add view more link
+                        f"🏷️ <a href='{item.url}'>{item.title}</a>\n"
+                        f"💰 Price: {item.price} {item.currency}\n"
+                        f"📍 Location: {item.location_country or 'N/A'}\n\n"
+                    )
+
                 print("About to send message: ")
                 notifier.send_message(message)
             return True
         except Exception as e:
             current_app.logger.error(f"Notification failed: {str(e)}")
             return False
-    
+
     @staticmethod
-    def send_test_notification(user, is_successfull_connection = False):
+    def send_test_notification(user, is_successfull_connection=False):
         try:
             chat_ids = [user.telegram_chat_ids['main']] + user.telegram_chat_ids['additional']
             for chat_id in chat_ids:
@@ -99,9 +93,8 @@ class NotificationManager:
                     notifier.send_message("TESTING TESTING 123...")
         except Exception as e:
             print(f"Error sending test notification: {str(e)}")
-            return False        
-    
-    
+            return False
+
     @staticmethod
     def send_price_drops(user, drops, query_text=None):
         chat_ids = [user.telegram_chat_ids['main']] + user.telegram_chat_ids['additional']
@@ -113,50 +106,45 @@ class NotificationManager:
             query_text = f" for '{query_text}'" if query_text else ""
             for drop in drops:
                 message = (
-                f"🛎️ **Price Alert{query_text}**\n"
-                f"📦 Item: {drop['item'].title}\n"
-                f"💰 Price dropped from £{drop['old_price']} → £{drop['new_price']}\n"
-                f"🔗 [View Item]({drop['item'].url})"
-            )
+                    f"🛎️ **Price Alert{query_text}**\n"
+                    f"📦 Item: {drop['item'].title}\n"
+                    f"💰 Price dropped from £{drop['old_price']} → £{drop['new_price']}\n"
+                    f"🔗 [View Item]({drop['item'].url})"
+                )
                 if user.notification_preferences.get('price_drops', True):
                     notifier.send_message(message)
-                
-    
+
     @staticmethod
     def send_auction_alerts(user, items, query_text=None):
         chat_ids = [user.telegram_chat_ids['main']] + user.telegram_chat_ids['additional']
         query_text = f" for '{query_text}'" if query_text else ""
-        
-        # Group items into batches of 5
+
         batch_size = 5
         item_batches = [items[i:i+batch_size] for i in range(0, len(items), batch_size)]
-        
+
         for chat_id in chat_ids:
             notifier = TelegramNotifier(
                 current_app.config['TELEGRAM_BOT_TOKEN'],
                 chat_id
             )
-            
+
             for batch in item_batches:
                 message = f"⏳ **Auctions Ending Soon{query_text}**\n\n"
-                
+
                 for idx, item in enumerate(batch, 1):
-                    # Process item data
                     current_bid = item.current_bid or item.price
                     item.end_time = item.end_time.replace(tzinfo=timezone.utc)
                     time_left = item.end_time - datetime.now(timezone.utc)
                     hours_left = round(time_left.total_seconds() / 3600, 1)
-                    
+
                     message += (
                         f"{idx}. 📦 **{item.title}**\n"
                         f"   💰 {item.current_bid} {item.current_bid_currency}\n"
                         f"   ⏰ {hours_left}h left\n"
                         f"   [View]({item.url})\n\n"
                     )
-                
-                # Add footer
+
                 message += f"Showing {len(batch)} of {len(items)} ending auctions"
-                
+
                 if user.notification_preferences.get('auction_alerts', True):
                     notifier.send_message(message)
-
