--- app/utils/scraper.py
@@ -1,3 +1,0 @@
-from app.searches.execution import create_ebay_client, scrape_ebay, scrape_new_items
-
-__all__ = ["create_ebay_client", "scrape_ebay", "scrape_new_items"]
